Add mention users feature for posts.
A user can be mentioned in posts by preceding the username with an @ symbol.
Upon posting, that user will receive an e-mail notification saying that
the user has been mentioned. The mention will also show up as a link when
viewing the post, so anyone can click the link to go to the mentioned user's
profile page.